Inject customer resource model as proxy into command class
When executing `bin/magento setup:install`, then Magento instantiates all the
registered Symfony console commands early in the process. At this point in time,
the database tables are not yet created. Therefore, if any resource models are
instantiated in the tree of object dependencies, then the installation crashes.
In order to prevent that "chain reaction of object instantiation" and enable
lazy-loading, such dependencies must be injected as proxies. With this change,
the `Customer` resource model that gets injected into the
`CustomerChangePasswordCommand` will only be instantiated when the command is
actually triggered, but no longer on Magento installation.